preliminary front end form for emails
